Moth aws methods to aws, fix tests
